requirements.txt
anyio==4.3.0
asgiref==3.8.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
nest-asyncio==1.6.0
oauthlib==3.2.2
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==21.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.4
Telethon==1.35.0
TgCrypto==1.2.5
tzdata==2024.1
urllib3==2.2.1
